created all 7 apis of category and generic product

+++ src/main/java/com/amitosh/productservice/Service/SelfProductService.java
@@ -3,7 +3,9 @@ package com.amitosh.productservice.Service;
 import com.amitosh.productservice.Model.Category;
 import com.amitosh.productservice.Model.Price;
 import com.amitosh.productservice.Model.Product;
+import com.amitosh.productservice.Repository.CategoryRepository;
 import com.amitosh.productservice.Repository.ProductRepository;
+import com.amitosh.productservice.dtos.CategoryDto;
 import com.amitosh.productservice.dtos.GenericProductDto;
 import com.amitosh.productservice.exceptions.NotFoundException;
 import org.springframework.context.annotation.Primary;
@@ -13,40 +15,46 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-@Primary
+
 @Service("selfProductService")
-public class SelfProductService implements productService {
+@Primary
+public class SelfProductService implements ProductService {
 
     private ProductRepository productRepository;
-    public SelfProductService(ProductRepository productRepository) {
+    private final CategoryRepository categoryRepository;
+
+    public SelfProductService(ProductRepository productRepository,
+                              CategoryRepository categoryRepository) {
         this.productRepository = productRepository;
+        this.categoryRepository = categoryRepository;
     }
 
     public GenericProductDto convertProductToGenericProduct(Product product) {
         GenericProductDto genericProductDto = new GenericProductDto();
         genericProductDto.setId(product.getId());
-        genericProductDto.setCategory(String.valueOf(product.getCategory()));
+        genericProductDto.setCategory(product.getCategory().getName());
         genericProductDto.setTitle(product.getTitle());
         genericProductDto.setDescription(product.getDescription());
         genericProductDto.setImage(product.getImage());
+        genericProductDto.setPrice(product.getPrice().getPrice());
         return genericProductDto;
     }
     @Override
     public GenericProductDto getProductById(UUID id) throws NotFoundException {
        Product product = this.productRepository.findById(id).get();
-       GenericProductDto genericProductDto = convertProductToGenericProduct(product);
-       return genericProductDto;
+        return convertProductToGenericProduct(product);
     }
 
     @Override
     public GenericProductDto createProduct(GenericProductDto genericProductDto) {
         Product product = new Product();
 
-        Category category = new Category();
+        Category category = categoryRepository.findDistinctByName(genericProductDto.getCategory());
         category.setName(genericProductDto.getCategory());
 
         Price price = new Price();
         price.setPrice(genericProductDto.getPrice());
+        price.setCurrency("Rupees");
 
         product.setPrice(price);
         product.setCategory(category);
@@ -54,6 +62,7 @@ public class SelfProductService implements productService {
         product.setId(genericProductDto.getId());
         product.setTitle(genericProductDto.getTitle());
         product.setImage(genericProductDto.getImage());
+
         Product savedProduct = this.productRepository.save(product);
         return convertProductToGenericProduct(savedProduct);
     }
@@ -70,26 +79,45 @@ public class SelfProductService implements productService {
 
     @Override
     public GenericProductDto deleteProduct(UUID id) throws NotFoundException {
-        Product product = this.productRepository.findById(id).get();
+        Optional<Product> product = this.productRepository.findById(id);
+        if (product.isEmpty()) {
+            throw new NotFoundException("product with id " + id + " not found");
+        }
         this.productRepository.deleteById(id);
-        return convertProductToGenericProduct(product);
+        return convertProductToGenericProduct(product.get());
     }
 
     @Override
-    public GenericProductDto update(GenericProductDto genericProductDto, UUID id) {
-        Product product = new Product();
-        Price price = new Price();
-        price.setPrice(genericProductDto.getPrice());
+    public GenericProductDto updateProduct(GenericProductDto genericProductDto, UUID id) throws NotFoundException {
+        Optional<Product> product = productRepository.findById(id);
+        if (product.isEmpty()) {
+            throw new NotFoundException("product with id "+ id +" not found");
+        }
+
+        Product updatedProduct = new Product();
         Category category = new Category();
         category.setName(genericProductDto.getCategory());
+        updatedProduct.setCategory(category);
+        updatedProduct.setTitle(genericProductDto.getTitle());
+        updatedProduct.setDescription(genericProductDto.getDescription());
+        updatedProduct.setImage(genericProductDto.getImage());
+        Price price = new Price("Rupees", genericProductDto.getPrice());
+        updatedProduct.setPrice(price);
+        Product product1 = productRepository.save(updatedProduct);
+        return convertProductToGenericProduct(product1);
+    }
 
-        product.setImage(genericProductDto.getImage());
-        product.setTitle(genericProductDto.getTitle());
-        product.setPrice(price);
-        product.setDescription(genericProductDto.getDescription());
-        product.setCategory(category);
-
-        this.productRepository.save(product);
-        return convertProductToGenericProduct(product);
+    @Override
+    public List<GenericProductDto> getAllProductsWithCategory(UUID category_id) throws NotFoundException {
+        List<GenericProductDto> productDtos = new ArrayList<>();
+        Optional<Category> category = categoryRepository.findById(category_id);
+        if (category.isEmpty()) {
+            throw new NotFoundException("category with id "+ category_id +" not found");
+        }
+        List<Product> products = productRepository.findAllByCategoryEquals(category.get());
+        for (Product product : products) {
+            productDtos.add(convertProductToGenericProduct(product));
+        }
+        return productDtos;
     }
 }
